requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
scikit-learn
numpy
pandas
xgboost
matplotlib
biopython
scikit-plot
imbalanced-learn
pyyaml